+++ hnsw/src/hnsw/params.rs
@@ -2,27 +2,27 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Params {
-    pub m: usize,
-    pub mmax: usize,
-    pub mmax0: usize,
+    pub m: u8,
+    pub mmax: u8,
+    pub mmax0: u8,
     pub ml: f32,
-    pub ef_cons: usize,
-    pub dim: usize,
+    pub ef_cons: u32,
+    pub dim: u32,
 }
 
 impl Params {
-    pub fn from_m(m: usize, dim: usize) -> Params {
+    pub fn from_m(m: u8, dim: u32) -> Params {
         Params {
             m,
             mmax: m,
-            mmax0: m * 2,
+            mmax0: (m * 2) as u8,
             ml: 1.0 / (m as f32).ln(),
-            ef_cons: m * 2,
+            ef_cons: (m as u32) * 2,
             dim,
         }
     }
 
-    pub fn from_m_efcons(m: usize, ef_cons: usize, dim: usize) -> Params {
+    pub fn from_m_efcons(m: u8, ef_cons: u32, dim: u32) -> Params {
         Params {
             m,
             mmax: m,
@@ -34,19 +34,19 @@ impl Params {
     }
 
     pub fn from(
-        m: usize,
-        ef_cons: Option<usize>,
-        mmax: Option<usize>,
-        mmax0: Option<usize>,
+        m: u8,
+        ef_cons: Option<u32>,
+        mmax: Option<u8>,
+        mmax0: Option<u8>,
         ml: Option<f32>,
-        dim: usize,
+        dim: u32,
     ) -> Params {
         Params {
             m,
             mmax: mmax.unwrap_or(m),
             mmax0: mmax0.unwrap_or(m * 2),
             ml: ml.unwrap_or(1.0 / (m as f32).ln()),
-            ef_cons: ef_cons.unwrap_or(m * 2),
+            ef_cons: ef_cons.unwrap_or((m as u32) * 2),
             dim,
         }
     }
